# packages/api-client/src/generated/modelli/aggiornaPreferenzeDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Prome API
 * Facciata REST versionata di Prome.
 * OpenAPI spec version: v1
 */

export interface AggiornaPreferenzeDto {
  commenti?: boolean;
  inviti?: boolean;
}